Refactor event handling and project operations in STOToolsKeybindManager
- Introduced a new `eventHandlers.js` module to encapsulate event listener setup, improving code organization and maintainability.
- Moved project-related operations to a new `projectOperations.js` module, enhancing separation of concerns and clarity in the codebase.
- Updated `app.js` to integrate the new modules, ensuring a cohesive user experience with improved event management and project handling.
- Removed redundant event listener code from `app.js`, streamlining the setup process and enhancing readability.

These changes significantly enhance the application's structure and maintainability, contributing to a more intuitive development experience.
